// SIGE/src/app/Empleados/interfaces/EmpleadoById.interfaces.ts
// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface EmpleadoByID {
  msj:     string;
  data:    DataById;
  errores: any[];
}

export interface DataById {
  edad:           number;
  idempleado:     number;
  identidad:      string;
  nombrecompleto: string;
  correo:         string;
  genero:         string;
  estado:         string;
  estcivil:       string;
  telefono:       string;
  tiposangre:     string;
  fechanac:       string;
  fechaIngreso:   string;
  imagen:         null;
  idusuario:      number;
  iddireccion:    number;
  idcargo:        number;
  idcontrato:     number;
  idgerencia:     number;
}